requirements.txt
annotated-types==0.6.0
anyio==4.3.0
astra-haystack==0.5.0
astrapy==0.7.7
backoff==2.2.1
boilerpy3==1.0.7
cassandra-driver==3.29.1
cassio==0.1.5
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
deprecation==2.1.0
distro==1.9.0
feedparser==6.0.11
filelock==3.13.1
fsspec==2024.3.1
geomet==0.2.1.post1
h11==0.14.0
h2==4.1.0
haystack-ai==2.0.0
haystack-bm25==1.0.2
hpack==4.0.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
hyperframe==6.0.1
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
lazy-imports==0.3.1
MarkupSafe==2.1.5
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
openai==1.14.2
packaging==24.0
pandas==2.2.1
pillow==10.2.0
posthog==3.5.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.6.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.4.0
tokenizers==0.15.2
toml==0.10.2
torch==2.2.1
tqdm==4.66.2
transformers==4.39.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
xmltodict==0.13.0
